requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorlog==6.7.0
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
jsonpath-ng==1.5.3
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pluggy==1.2.0
ply==3.11
pytest==7.4.1
pytest-instafail==0.5.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
tomli==2.0.1
tzdata==2023.3
urllib3==2.0.3
xlrd==2.0.1
XlsxWriter==3.1.2

xlwt~=1.3.0